% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_calculations.R
\name{AICcalc}
\alias{AICcalc}
\title{Calculate AICc parameter for models}
\usage{
AICcalc(RSS, nPC, nvert, noregions)
}
\arguments{
\item{RSS}{total residual sums of squares, summed accross regions and variables}

\item{nPC}{Number of PCs analyzed}

\item{nvert}{Number of vertebrae}

\item{noregions}{Number of regions}
}
\value{
AICc
}
\description{
Calculates corrected AIC values for least squares regression using residuals
}
\details{
K is 2 (slope, intercept) multiplied by number of variables and number of regions

No vert must be > no regions*3
}
