% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-all_valid_nuts_codes.R
\docType{data}
\name{all_valid_nuts_codes}
\alias{all_valid_nuts_codes}
\title{European Union: All Valid NUTS Codes}
\format{
A data frame with 3 variables:
\describe{
\item{geo}{NUTS geo identifier}
\item{typology}{country, NUTS1, NUTS2 or NUTS3}
\item{nuts}{The NUTS definition where the geo code can be found.}
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/nuts/history/}
}
\usage{
all_valid_nuts_codes
}
\description{
A dataset containing all recognised geo codes in the EU
NUTS correspondence tables. This is re-arranged from
\code{\link{nuts_changes}}.
}
\seealso{
nuts_recoded, nuts_changes
}
\keyword{datasets}
