% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alligator}
\alias{alligator}
\title{Measurements from the vertebral column of Alligator}
\format{A matrix with 22 vertebrae and 19 measurements. Column 1, vertebra, is the positional
information.}
\usage{
alligator
}
\description{
Linear and angular measurements from Alligator mississipiensis MCZ 81457
}
\keyword{datasets}
