% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{angtransform}
\alias{angtransform}
\title{Cos transforms angles}
\usage{
angtransform(data, angles)
}
\arguments{
\item{data}{Raw data including angles in degrees.}

\item{angles}{Vector of strings indicating columns which contain angles.}
}
\description{
Converts angles to radians then cos transforms them.
}
