% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-federal_states.R
\docType{data}
\name{australia_states}
\alias{australia_states}
\title{Australia: States And Territories}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{country_code}{ISO 3166-1 country codes}
\item{geo_code}{subdivision codes within Australia (states and territories)}
\item{geo_name}{subdivision names within Australia (states and territories)}
}
}
\source{
The Online Browsing Platform of the International Organization for
Standardization \url{https://www.iso.org/obp/ui/#iso:code:3166:AU}
}
\usage{
australia_states
}
\description{
A dataset containing the states and territories of Australia.
}
\keyword{datasets}
