% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{axesplot}
\alias{axesplot}
\title{Plot PCO axes}
\usage{
axesplot(data, x, y, Xvar)
}
\arguments{
\item{data}{PCO scores or any multivariate data matrix.}

\item{x}{Variable for x axis.}

\item{y}{Variable for y axis.}

\item{Xvar}{Variable for labelling points, usually positional variable such as vertebral count.}
}
\description{
Compare any PCO axes along the series.
}
