% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regions.R
\name{calcregions}
\alias{calcregions}
\title{Calculate segmented regression models}
\usage{
calcregions(Xvar, Yvar, noregions)
}
\arguments{
\item{Xvar}{Positional variable e.g., vertebral count.}

\item{Yvar}{Dependent variable.}

\item{noregions}{Maximum number of regions to be computed, up to 6.}
}
\value{
Returns a matrix containing residual sums of squares and r-squared values for each model.
}
\description{
\code{calcregions} fits segmented regression models to all possible regions, up to noregions number of regions.
}
