% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_regions.R
\name{compileregions}
\alias{compileregions}
\title{Calculates region model fit}
\usage{
compileregions(Xvar, data, noregions = 2)
}
\arguments{
\item{Xvar}{Positional variable e.g., vertebral count.}

\item{data}{Matrix of dependent variables.}

\item{noregions}{Maximum number of regions to be computed, up to 6.}
}
\value{
Returns a data matrix with one row for each model. Columns include model parameters
 (breakpoints) and residual sums of squares for each dependent variable
}
\description{
Uses \code{calcregions} to calculate region models for each variable separately, then
exports the summed residual sums of squares.
}
\details{
This function feeds data into \code{calcregions}, compiles the results and calculates model fitting
metrics.
}
\examples{
data("alligator")
Xvar=alligator[,1]
data=alligator[,2:ncol(alligator)]
regiondata=compileregions(Xvar, data, 2)
}
