% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{confintModel}
\alias{confintModel}
\title{Calculate confidence interval on model based on simulation}
\usage{
confintModel(tree, testsim, testmod, reps, dat)
}
\arguments{
\item{tree}{tree}

\item{testsim}{simulation function}

\item{testmod}{model for testing}

\item{reps}{no of simulations}

\item{dat}{no of simulations}
}
\value{
median and confidence interval on simulations
}
\description{
Estimate confidence intervals by simulating data based on true OU model. Based on Boettiger et al. (2011)
}
