% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_nuts_lau_2019}
\alias{create_nuts_lau_2019}
\title{Create the nuts_lau_2019 correspondence table
May be used to create similar historical correspondence tables.}
\usage{
create_nuts_lau_2019()
}
\value{
A data.frame which is also saved and can be retrieved with
\code{data(nuts_lau_2019).} Use this function as a template to
obtain historical correspondence tables.
}
\description{
Create the nuts_lau_2019 correspondence table
May be used to create similar historical correspondence tables.
}
