% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-daily_internet.R
\docType{data}
\name{daily_internet_users}
\alias{daily_internet_users}
\title{Daily Internet Users}
\format{
A data frame with 3 variables:
\describe{
\item{geo}{National and sub-national geographical codes from Eurostat}
\item{time}{Time, coded as a numeric variable of the year, 2006-2019}
\item{values}{The numeric statistical values}
}
}
\usage{
daily_internet_users
}
\description{
A dataset containing the percentage of individuals who used the Internet on
a daily basis in the European countries and regions.
}
\details{
The fresh version of this statistic can be obtained by
\code{eurostat::get_eurostat("isoc_r_iuse_i", time_format = "num")}
and filtered for the \code{indic_is = "I_IDAY"} indicator and the
\code{unit="PC_IND"} unit.
}
\keyword{datasets}
