% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_functions.R
\name{disparityregions}
\alias{disparityregions}
\title{Calculate disparity between regions}
\usage{
disparityregions(data, bestmodel)
}
\arguments{
\item{data}{data frame with positional information in column "vert"}

\item{bestmodel}{vector in format (no regions, break 1, break 2...)}
}
\value{
var Variance of means

mean Region means
}
\description{
Calculates variance of region means based on pre-defined region model
}
