% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_selection.R
\name{eigenplot}
\alias{eigenplot}
\title{Plot eigenvalues}
\usage{
eigenplot(eigen.true, eigen.boot)
}
\arguments{
\item{eigen.true}{Vector of actual eigenvectors.}

\item{eigen.boot}{Matrix of bootstrapped eigenvectors.}
}
\description{
Produces a scree plot of eigenvalues from a PCO and compares them with eigenvalues of
randomized data, obtained from \code{PCOcutoff}.
}
