% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{islandscheck}
\alias{islandscheck}
\title{Check for islands in surface analysis}
\usage{
islandscheck(tree, reps, dat, credthresh = 5, samplethresh = 5)
}
\arguments{
\item{tree}{tree}

\item{reps}{no of repeats}

\item{dat}{data}

\item{credthresh}{AIC cutoff for credible models. Default is 5.}

\item{samplethresh}{AIC cutoff for sampling shifts. Default is 5.}
}
\value{
"Real" model plus stats on variation within resampled models
}
\description{
Rerun surface analysis with sampling of shifts. Select only credible models based on AIC threshold.
}
