% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{Missingval}
\alias{Missingval}
\title{Estimate and fill missing values}
\usage{
Missingval(data)
}
\arguments{
\item{data}{data with missing elements}
}
\value{
data Data with missing values filled
}
\description{
Estimates missing values by averaging values from nearest adjacent vertebrae.
}
\details{
In the case of a first or last vertebra, the nearest value is taken. Will only fill up to two missing data points.
}
