% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mixed_nuts_example.R
\docType{data}
\name{mixed_nuts_example}
\alias{mixed_nuts_example}
\title{Example Data Frame: Mixed EU Typologies.}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{geo}{NUTS geo identifier, mixed from 4 typology levels.}
\item{values}{Random numbers.}
\item{method}{Descriptive metadata.}
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/nuts/history/}
}
\usage{
mixed_nuts_example
}
\description{
This data frame is a fictious example that contains in a small,
easy-to-review example many potential typological problems. It is
used to test imputation functions and to create examples with them.
}
\seealso{
nuts_changes, all_valid_nuts_codes, impute_down_nuts
}
\keyword{datasets}
