% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_calculations.R
\name{model_support}
\alias{model_support}
\title{Calculate relative support for each model}
\usage{
model_support(models, nvert, nPC)
}
\arguments{
\item{models}{Data frame containing models and summed RSS}

\item{nvert}{No of vertebrae}

\item{nPC}{No of PCs used}
}
\value{
Model_support Data frame containing each model and support values inlcuding AICc and Akaike weights

Region_score Mean weighted region score
}
\description{
Calculates AIC ratios and Akaike weights for each model to determine its strength of support.
}
\details{
Also calculates a weighted average region score, which reflects the probability of each region model
}
