% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_calculations.R
\name{modelselect}
\alias{modelselect}
\title{Select best region models from segmented regressions}
\usage{
modelselect(regiondata, noregions, nopcos, startpco = 1)
}
\arguments{
\item{regiondata}{Matrix created by \code{compileregions} containing region breaks and summed residuals}

\item{noregions}{Number of regions}

\item{nopcos}{Number of PCOs to be analyzed}

\item{startpco}{The pco to begin selection with}
}
\value{
Data frame of best models for given regions
}
\description{
Select the best model for each number of regions by minimizing the residual sums of squares.
}
