% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_calculations.R
\name{multvarrsq}
\alias{multvarrsq}
\title{Calculate multivariate adjusted r squared}
\usage{
multvarrsq(Xvar, data, modelsupport)
}
\arguments{
\item{Xvar}{Independent variable}

\item{data}{Multiple PCO/PCA score}

\item{modelsupport}{Model_support object from \code{model_support}.}
}
\value{
rsq multivariate uncorrected r-squared

adjrsq multivariate adjusted r-squared
}
\description{
Calculate rsqaured based on sums of squares from across multiple PCOs.
}
\details{
Sum up sums of squares and degrees of freedom from separate PCO fits to estimate overall fit of data
}
