% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{musm}
\alias{musm}
\title{Measurements from the vertebral column of Mus}
\format{A matrix with 23 vertebrae and 19 measurements. Column 1, vertebra, is the positional
information.}
\usage{
musm
}
\description{
Linear and angular measurements from Mus musculus MCZ 59560
}
\keyword{datasets}
