% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuts_exceptions.R
\docType{data}
\name{nuts_exceptions}
\alias{nuts_exceptions}
\title{NUTS Coding Exceptions}
\format{
A data frame with 2 variables:
\describe{
\item{geo}{National and sub-national geographical codes from Eurostat}
\item{typology}{Short description of exception}
}
}
\source{
Eurostat NUTS history: \url{https://ec.europa.eu/eurostat/web/nuts/history/}
}
\usage{
nuts_exceptions
}
\description{
A dataset containing exceptions to the NUTS geographical codes.
}
\details{
They contains non-EU
regions that are consistent with NUTS, but not defined within the NUTS.

The also contain European country codes that do not conform with NUTS.
}
\seealso{
nuts_recoded, nuts_changes, all_valid_nuts_codes
}
\keyword{datasets}
