% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eu_lau_2019.R, R/data-nuts_lau_2019.R
\docType{data}
\name{nuts_lau_2019}
\alias{nuts_lau_2019}
\title{European Union: NUTS And LAU Correspondence}
\format{
A data frame with 398590 rows and 20 variables:
\describe{
\item{nuts_3_code}{NUTS3 code of the local administrative unit}
\item{lau_code}{The year when the code was first used}
\item{lau_name_national}{The year when the code was last used}
\item{lau_name_latin}{NUTS code in the 2003 definition}
\item{change_y_n}{NUTS code in the 2003 definition}
\item{population}{NUTS code in the 2006 definition}
\item{total_area_m_2}{NUTS code in the 2010 definition}
\item{degurba}{Degree of urbanization}
\item{geg_change_compared_to_last_year}{NUTS code in the 2016 definition}
\item{coastal_area_yes_no}{NUTS code in the 2021 definition}
\item{coast_change_compared_to_last_year}{NUTS territorial name in the 2003 definition}
\item{city_id}{NUTS territorial name in the 2006 definition}
\item{city_id_change_compared_to_last_year}{NUTS territorial name in the 2010 definition}
\item{city_name}{NUTS territorial name in the 2013 definition}
\item{greater_city_id}{NUTS territorial name in the 2016 definition}
\item{greater_city_id_change_compared_to_last_year}{NUTS territorial name in the 2021 definition}
\item{greater_city_name}{Change described in the 2003 correspondence table}
\item{fua_id}{Change described in the 2006 correspondence table}
\item{fua_id_change_compared_to_last_year}{Change described in the 2010 correspondence table}
\item{fua_name}{Change described in the 2013 correspondence table}
}

A data frame with 99140 rows and 22 variables:
\describe{
\item{nuts_3_code}{NUTS3 code of the local administrative unit, 2016 definition}
\item{lau_code}{Local Administrative Unit code}
\item{lau_name_national}{LAU name, official in national language(s)}
\item{lau_name_latin}{LAU name, official Latin alphabet version}
\item{name_change_last_year}{Change in name in the year before?}
\item{population}{Population}
\item{total_area_m2}{Area in square meters}
\item{degurba}{Degree of urbanization}
\item{degurba_change_last_year}{Change in degree of urbanization?}
\item{coastal_area}{Part of coastal area classification?}
\item{coastal_change_last_year}{Change in coastal area classification}
\item{city_id}{NUTS territorial name in the 2006 definition}
\item{city_id_change_last_year}{NUTS territorial name in the 2010 definition}
\item{city_name}{Name of the city}
\item{greater_city_id}{Containing metro area ID, if applicable}
\item{greater_city_id_change_last_year}{Change in metro area ID}
\item{greater_city_name}{Name of containing greater city (metropolitan) area, if applicable}
\item{fua_id}{FUA ID}
\item{fua_id_change_last_year}{Change of FUA ID since last year}
\item{fua_name}{Name in FUA database}
\item{country}{NUTS country code (EL for Greece, UK for United Kingdom)}
\item{gisco_id}
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/nuts/local-administrative-units}

\url{https://ec.europa.eu/eurostat/web/nuts/local-administrative-units}
}
\usage{
nuts_lau_2019

nuts_lau_2019
}
\description{
A dataset containing the joined correspondence tables of the
EU NUTS and local administration units (LAU) typologies.

A dataset containing the joined correspondence tables of the
EU NUTS and local administration units (LAU) typologies.
}
\details{
This is also the authoritative vocabulary for local administration,
names, including city and metropolitan area names.
}
\seealso{
nuts_recoded, all_valid_nuts_codes

nuts_recoded, all_valid_nuts_codes
}
\keyword{datasets}
