% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{pairwiseLoglik}
\alias{pairwiseLoglik}
\title{Calculate pairwise loglikelihood of models}
\usage{
pairwiseLoglik(tree, sim1, sim2, lik1, lik2, reps, dat)
}
\arguments{
\item{tree}{tree}

\item{sim1}{function for simulating data under model1}

\item{sim2}{function for simulating data under model2}

\item{lik1}{function for fitting model1}

\item{lik2}{function for fitting model2}

\item{reps}{no of simulations for each model}

\item{dat}{the 'real' data for calculating real log lik}
}
\value{
a plot comparing distributions to real ration
}
\description{
Test evolutionary hypotheses by comparing the distribution of their log-liklihood ratios
}
\details{
Based on Boettiger et al. (2011). First likelihood ratios are generated based on data simulated under model 1, comparing
liklihood of model1 and model2. Then data are simulated under model2 and the liklihood ratios
are computed again. By comparing the distribution of these ratios with the 'true' ratio
the significance of the liklihood can be tested.

Formatted to interact primarily with "Surface" models using 'surfaceSimulate' and 'startingModel'
}
