% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pco.load}
\alias{pco.load}
\title{Estimate loadings on PCOS}
\usage{
pco.load(data, PCOscore)
}
\arguments{
\item{data}{Raw data}

\item{PCOscore}{PCO scores}
}
\value{
loadings correlation coefficients and direction of correlation
}
\description{
Estimates loadings based on correlation coefficients of PCOs against raw data
}
