% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_selection.R
\name{PCOcutoff}
\alias{PCOcutoff}
\title{Find signficant PCOs}
\usage{
PCOcutoff(data, nreps, metric)
}
\arguments{
\item{data}{Raw data.}

\item{nreps}{Number of iterations.}

\item{metric}{Distance matrix calculation metric. "euclidean", "manhattan", or "gower".}
}
\value{
Eigenvalues of original PCO (eigen.true), mean eigenvalues from the randomized data (eigen.mean),
standard deviation of randomized eigenvalues (eigen.sd), PCOs with eigenvalues exceeding the
randomized data (sigpco), matrix of all eigenvalues produced by randomization (eigen.boot).
}
\description{
Compares eigenvalue distribution of PCO to that with randomized data in order to extract
axes with significant signal.
}
\details{
Significant PCO's are defined as those with greater eigenvalues than from random data. Therefore
the PCO cutoff is axes whose eigenvalues fall below the mean eigenvalue for that axis from the
randomized data. Data are randomly sampled by row.

Warning: Bootstrapping is sensitive to unequale variances of columns. You may want to use
\code{scale} to normalize before using this approach
}
