% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_selection.R
\name{PCOmax}
\alias{PCOmax}
\title{Find the number of PCOs which yeilds the Maximum regionscore}
\usage{
PCOmax(regiondata, noregions, nvert)
}
\arguments{
\item{regiondata}{Object returned from \code{compileregions}}

\item{noregions}{Maximum number of regions}

\item{nvert}{Number of vertebrae}
}
\value{
pco.max No of PCOs producing the largest regionscore

pco.dist Region score for each cumulative addition of PCOs
}
\description{
Searches top 10 PCOs to find the peak of regionscore
}
