% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator
See \code{magrittr::\link[magrittr]{\%>\%}} for details.}
\description{
Pipe operator
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\keyword{internal}
