% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.pco.reg}
\alias{plot.pco.reg}
\title{Plot the scree distribution of regionscores across pcos}
\usage{
\method{plot}{pco.reg}(eigenvals, nvert, namelabel, regiondata, noregions)
}
\arguments{
\item{eigenvals}{Eigenvalues from PCO}

\item{nvert}{Number of vertebrae}

\item{namelabel}{Label for plot (specimen name)}

\item{regiondata}{Segmented regression models for specimen from \code{compileregions}}

\item{noregions}{Maximum number of regions}
}
\description{
Plot indicating how regionscore is influenced by including more PCOs
}
\details{
Plots cumulative variance (line), regionscore based on each separate PCO (black) and regionscore based
on cumulative PCOs (red)
}
