% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotsegreg}
\alias{plotsegreg}
\title{Plot best segmented regression model}
\usage{
plotsegreg(Xvar, pcono, data, modelsupport)
}
\arguments{
\item{Xvar}{Positional vector such as vertebral count.}

\item{pcono}{Which PC to plot}

\item{data}{PCO scores.}

\item{modelsupport}{Model_support object from \code{model_support}.}
}
\description{
Plots the segmented regression for the top model on PCO1.
}
