% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotvar}
\alias{plotvar}
\title{Make a scree plot of variables against species}
\usage{
plotvar(yvar, labels, header)
}
\arguments{
\item{yvar}{yvar}

\item{labels}{labels}

\item{header}{header}
}
\description{
Make a scree plot of variables against species
}
