% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_calculations.R
\name{region_anal}
\alias{region_anal}
\title{Run regionalization analysis}
\usage{
region_anal(Xvar, scores, noregions, nopcos)
}
\arguments{
\item{Xvar}{Vector, x variable}

\item{scores}{Matrix, PCO scores}

\item{noregions}{Number of region}

\item{nopcos}{Number of PCOs}
}
\value{
Model_support support for each best model

Region_scores Weighted region scores
}
\description{
Wrapper for whole regionalization analysis, inlcuding running the models,
selecting the models and calculating the model support
}
