% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{regionheat}
\alias{regionheat}
\title{Plots heatmap of region boundaries}
\usage{
regionheat(regiondata, noregions, nopcos, Xvar)
}
\arguments{
\item{regiondata}{region data}

\item{noregions}{number of regions for which to generate heat score}

\item{nopcos}{No of PCOs}

\item{Xvar}{Positional variable}
}
\value{
heatscore
}
\description{
Relative probability of break and each position based on RSS for a given number of regions
}
