% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_calculations.R
\name{regionrsq}
\alias{regionrsq}
\title{R-squared for Best Model}
\usage{
regionrsq(Xvar, data, pcono, modelsupport)
}
\arguments{
\item{Xvar}{X variable}

\item{data}{PCO scores}

\item{pcono}{Which column of data to use}

\item{modelsupport}{Model_support object from \code{model_support}.}
}
\value{
adjrsq adjusted r-squared

SStot Total sums of squares

df Total degrees of freedom

SSres Residual sums of squares

dfe Error degrees of freedom
}
\description{
Calculate the fit of a segmented regression model univariately
}
