% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analyses.R
\name{sens_anal}
\alias{sens_anal}
\title{Run Sensistivity analysis}
\usage{
sens_anal(Xvar, data, noregions, nopcos, missing, type, reps,
  fillmiss = TRUE)
}
\arguments{
\item{Xvar}{Positional variable}

\item{data}{data}

\item{noregions}{Number of regions}

\item{nopcos}{Number of PCOs}

\item{missing}{proportion missing, between zero and one}

\item{type}{"elem", "vert", or "var"}

\item{reps}{Number of repeats}

\item{fillmiss}{Fill missing TRUE or FALSE}
}
\value{
regionscore

AIC models
}
\description{
Reruns regionalization analysis on bootstrapped data
}
\details{
Data will be bootstrapped by element (cells in matrix), by vertebra (rows), or by variable (cols)
}
