% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analyses.R
\name{sim_missing}
\alias{sim_missing}
\title{Simulate missing data}
\usage{
sim_missing(data, missing, type, Xvar)
}
\arguments{
\item{data}{data matrix}

\item{missing}{proportion missing data, from zero to one}

\item{type}{"elem", "vert", "var" randomly replace elements, whole rows, or whole columns}

\item{Xvar}{Independent variable}
}
\value{
data

xvar
}
\description{
Randomly replace data elements with NAs, or remove whole vertebrae or variables
}
\details{
If vertebrae are removed Xvar is corrected
}
