% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{single.model}
\alias{single.model}
\title{Calculate single region model}
\usage{
single.model(Xvar, data, noregions, breaks)
}
\arguments{
\item{Xvar}{Independent variable}

\item{data}{Matrix of dependent variables}

\item{noregions}{Number of regions}

\item{breaks}{Vector of breakpoints}
}
\value{
RSS.results residuals sums of squares for each PC and each region
}
\description{
Code for examining the residuals for a single region model with specified parameters
}
\details{
Useful for examining the fit of a specified region model
}
