% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_selection.R
\name{svdPCO}
\alias{svdPCO}
\title{Calculate PCO (principal co-ordinates analysis) based on svd}
\usage{
svdPCO(x, metric)
}
\arguments{
\item{x}{Raw data matrix.}

\item{metric}{Distance matrix calculation metric. "euclidean", "manhattan", or "gower".}
}
\value{
PCO scores (scores) and eigenvalues (eigen.val)
}
\description{
Calculates distance matrix from raw data, then conducts a PCO ordination using a
single value decomposition.
}
\details{
This differs from other PCO functions which use \code{CMDscale} that depends on a
spectral decomposition.
}
