% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nuts.R
\name{validate_nuts_country}
\alias{validate_nuts_country}
\title{Validate comformity with NUTS country codes}
\usage{
validate_nuts_country(dat, geo_var = "geo")
}
\arguments{
\item{dat}{A data frame with a 2-character geo variable to be validated}

\item{geo_var}{Defaults to \code{"geo"}. The variable that contains the
2 character geo codes to be validated.}
}
\description{
Validate comformity with NUTS country codes
}
\examples{
{
my_dat <- data.frame ( 
 geo = c("AL", "GR", "XK", "EL", "UK", "GB", "NLD", "ZZ" ), 
 values = runif(8)
 )
 validate_nuts_country(my_dat)
}

}
