% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_nuts_regions.R
\name{validate_nuts_region}
\alias{validate_nuts_region}
\title{Validate Comformity With NUTS Geo Codes}
\usage{
validate_nuts_region(dat, geo_var = "geo", nuts_year = 2016)
}
\arguments{
\item{dat}{A data frame with a 3-5 character \code{geo_var} variable
to be validated.}

\item{geo_var}{Defaults to \code{"geo"}. The variable that contains the
3-5 character geo codes to be validated.}
}
\description{
Validate Comformity With NUTS Geo Codes
}
\examples{
{
my_reg_data <- data.frame ( 
  geo = c("BE1", "HU102", "FR1", 
          "DED", "FR7", "TR", "DED2",
          "EL", "XK", "GB"), 
  values = runif(10))

validate_nuts_region (my_reg_data)
}
}
\seealso{
Other validate functions: 
\code{\link{validate_nuts_country}()}
}
\concept{validate functions}
