% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nuts.R
\name{validate_nuts_regions}
\alias{validate_nuts_regions}
\title{Validate comformity with NUTS geo codes}
\usage{
validate_nuts_regions(dat, geo_var = "geo", nuts_year = 2016)
}
\arguments{
\item{dat}{A data frame with a 3-5 character geo variable to be validated}

\item{geo_var}{Defaults to \code{"geo"}. The variable that contains the
3-5 character geo codes to be validated.}
}
\description{
Validate comformity with NUTS geo codes
}
\examples{
{
my_reg_data <- data.frame ( 
  geo = c("BE1", "HU102", "FR1", 
          "DED", "FR7", "TR", "DED2",
          "EL", "XK", "GB"), 
  values = runif(10))

validate_regional_df ( my_reg_data )
}
}
