% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_param}
\alias{validate_param}
\title{Validate Mandatory Parameters}
\usage{
validate_param(param, param_name)
}
\arguments{
\item{param}{A parameter value that must not be \code{NULL}.}

\item{param_name}{The name of the parameter that must not have a value of \code{NULL}.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
These parameters must not be \code{NULL}. The \code{param_name} is needed for a
meaningful error message.
}
\keyword{internal}
