% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_typology}
\alias{validate_typology}
\title{Validate typology Parameter}
\usage{
validate_typology(typology)
}
\arguments{
\item{typology}{Currently the following typologies are supported:
\code{"NUTS1"}, \code{"NUTS2"}, \code{"NUTS3"} or \code{"NUTS"} for any of
the NUTS typologies. The technical typology \code{"NUTS0"}
can be used to translate Eurostat country codes to ISO 3166-1 alpha-2
country codes.}
}
\value{
A boolean, logical variable if the typology in question exists, the typology parameter is valid.
}
\description{
Validate typology Parameter
}
\keyword{internal}
