% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_nuts.R
\name{recode_nuts}
\alias{recode_nuts}
\title{Recode Region Codes From Source To Target NUTS Typology}
\usage{
recode_nuts(
  dat,
  geo_var = "geo",
  source_nuts_year = 2016,
  target_nuts_year = 2013
)
}
\arguments{
\item{dat}{A data frame with a 3-5 character \code{geo_var} variable
to be validated.}

\item{geo_var}{Defaults to \code{"geo"}. The variable that contains the
3-5 character geo codes to be validated.}

\item{source_nuts_year}{Defaults to 2016}

\item{target_nuts_year}{Defaults to 2013}
}
\description{
Recode Region Codes From Source To Target NUTS Typology
}
