% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{regionheat}
\alias{regionheat}
\title{Plots heatmap of region boundaries}
\usage{
regionheat(regiondata, noregions, nopcos, Xvar)
}
\arguments{
\item{regiondata}{region data}

\item{noregions}{number of regions for which to generate heat score}

\item{nopcos}{No of PCOs}

\item{Xvar}{Positional variable}
}
\value{
heatscore- probability of breaks at a given poistion

stdev - standard deviation of each breakpoint
}
\description{
Relative probability of break and each position based on RSS for a given number of regions
}
\details{
Also calculates the standard deviation of each breakpoint in the top 5% of models based on RSS
If there are fewer then five models in the top 5%, the top five are used.
}
