% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{package}
\name{regions}
\alias{regions}
\title{regions: A package for working with regional statistics.}
\description{
The foo package provides three categories of important functions:
validate, impute and aggregate.
}
\section{validate functions}{

The validate functions validate the comformity of a typological
(geographical) label with a certain typology.  Currently the EU statistical
NUTS typologies and countries are implemented.
}

\section{impute functions}{

The impute functions impute data from one regional unit to a different
level of regional unit, such as a country level data to a province / state
level data.
}

