% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Validate Parameter 'dat'}
\usage{
validate_data_frame(dat, geo_var = NULL, nuts_year = NULL)
}
\arguments{
\item{dat}{A data frame input to be validated.}

\item{geo_var}{The variable that contains the geographical codes in the
NUTS typologies.}

\item{nuts_year}{The year of the NUTS typology to use.}
}
\description{
Validate Parameter 'dat'
}
