% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Validate Parameter 'dat'}
\usage{
validate_data_frame(
  dat,
  geo_var = NULL,
  nuts_year = NULL,
  values_var = NULL,
  method_var = NULL
)
}
\arguments{
\item{dat}{A data frame input to be validated.}

\item{geo_var}{The variable that contains the geographical codes in the
NUTS typologies, defaults to code{"geo_var".}}

\item{nuts_year}{The year of the NUTS typology to use.}

\item{values_var}{The variable that contains the upstream data to be
imputed to the downstream data, defaults to \code{"values"}.}

\item{method_var}{The variable that contains the metadata on various
processing information, defaults to \code{NULL} in which case it will
be returned as \code{'method'}.}
}
\description{
Validate Parameter 'dat'
}
\keyword{internal}
