% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Validate Parameter 'dat'}
\usage{
validate_data_frame(dat)
}
\arguments{
\item{dat}{A data frame input to be validated.}
}
\description{
Validate Parameter 'dat'
}
